Move ./src/components/Projects to ./src/projects
