+++ src/App.tsx
@@ -3,9 +3,9 @@ import "primereact/resources/themes/bootstrap4-light-blue/theme.css";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { Header } from "./components/Header";
 import Help from "./components/Help";
-import { globaliseConfig } from "./components/Projects/globalise/config/";
-import { mondriaanConfig } from "./components/Projects/mondriaan/config";
-import { republicConfig } from "./components/Projects/republic/config";
+import { globaliseConfig } from "./projects/globalise/config/";
+import { mondriaanConfig } from "./projects/mondriaan/config";
+import { republicConfig } from "./projects/republic/config";
 import { Search } from "./components/Search/Search";
 import { Detail } from "./Detail";
 import { ErrorPage } from "./ErrorPage";
