+++ src/projects/default/config/englishLabels.ts
@@ -1,4 +1,4 @@
-import {Labels} from '../../../../model/Labels.tsx';
+import {Labels} from '../../../model/Labels.tsx';
 
 export const englishLabels: Labels = {
   DATE_ASC: 'Date (ascending)',
