+++ src/projects/globalise/RenderMetadataPanel.tsx
@@ -1,5 +1,5 @@
 import { useParams } from "react-router-dom";
-import { AnnoRepoAnnotation } from "../../../model/AnnoRepoAnnotation";
+import { AnnoRepoAnnotation } from "../../model/AnnoRepoAnnotation";
 
 type RenderMetadataPanelProps = {
   annotations: AnnoRepoAnnotation[];
