+++ src/projects/globalise/getAnnotationItem.ts
@@ -1,4 +1,4 @@
-import { AnnoRepoAnnotation } from "../../../model/AnnoRepoAnnotation";
+import { AnnoRepoAnnotation } from "../../model/AnnoRepoAnnotation";
 
 export const getAnnotationItem = (annotation: AnnoRepoAnnotation) => {
   return annotation.body.type;
