+++ src/projects/mondriaan/RenderSearchInfoPage.tsx
